LayoutTests/editing/selection/5354455-1.html

<script src="../../resources/dump-as-markup.js"></script>

<p>This tests whether right clicking on a paragraph break in editable content selects it.  The break should be selected on OS X, but not on Windows or Linux.  To run it manually, right click on the paragraph break after the first paragraph below.</p>
<div id="dump">
<div id="div" contenteditable="true">
<div><span id="text">The following paragraph break should be selected on OS X.</span></div>
<br>
</div>
</div>

<script>
if (window.layoutTestController) {
    paragraph = document.getElementById("text");
    x = paragraph.offsetParent.offsetLeft + paragraph.offsetLeft + paragraph.offsetWidth + 10;
    y = paragraph.offsetParent.offsetTop + paragraph.offsetTop + paragraph.offsetHeight / 2;
    eventSender.mouseMoveTo(x, y);
    // Give the editable region focus.
    eventSender.mouseDown();
    eventSender.mouseUp();
    // Right click on the paragraph break to select it.
    eventSender.contextClick();
}
Markup.dump(document.getElementById("dump"));
</script>
