LayoutTests/platform/mac/editing/selection/5354455-1-expected.txt

| "
"
| <div>
|   contenteditable="true"
|   id="div"
|   "
"
|   <div>
|     <span>
|       id="text"
|       "The following paragraph break should be selected on OS X.<#selection-anchor>"
|   "
"
|   <#selection-focus>
|   <br>
|   "
"
| "
"
